app/requirements.txt
streamlit==1.28.0
numpy==1.24.4
matplotlib==3.7.0
pandas==2.1.1
scanpy==1.9.5
torch==2.1.0
torchvision==0.16.0
torch_geometric==2.4.0
tqdm==4.64.1
scikit-learn==1.2.1
scikit-misc==0.3.0
scipy==1.10.1
seaborn==0.12.2
umap-learn==0.5.4
streamlit_lottie==0.0.5
psycopg2-binary==2.9.9
sqlalchemy==2.0.22
leidenalg==0.9.1
anndata==0.10.2
scvi-tools==1.0.3
squidpy==1.3.1
bokeh==2.3.1
h5py==3.8.0
imageio==2.9.0   
python-igraph==0.10.4
python-louvain==0.15 
scikit-network==0.28.3
seaborn==0.12.2
psutil==5.9.4
pydantic==1.10.8
louvain==0.8.1
python-dotenv==1.0.0
scrublet==0.2.3
pytest==7.4.3